print the first 98 fibonacci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,13 +7,13 @@
 
 int main(void)
 {
-	unsigned int num1 = 1;
-	unsigned int num2 = 2;
-	unsigned int i;
-	unsigned int next;
+	unsigned long int num1 = 1;
+	unsigned long int num2 = 2;
+	unsigned long int i;
+	unsigned long int next;
 
 	printf("%lu,%lu, " num1, num2);
-	
+
 	for (i = 0; i < 96; i++)
 	{
 		printf("%lu, ", next);
